pyproject.toml
[project]
name = "pytest-testrail-plugin"
version = "0.1.0"
description = "A Pytest plugin for real-time reporting of test results to TestRail"
license = "MIT"
authors = [
    {name = "Alok Nag",email = "nag.alok@gmail.com"}
]
readme = "README.md"
requires-python =  ">=3.13,<4.0"
dependencies = [
    "pytest (>=8.3.5,<9.0.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]

[tool.poetry]
packages = [{include = "pytest_testrail_plugin", from = "src"}]

[tool.poetry.plugins.pytest11]
pytest-testrail = "pytest_testrail_plugin.plugin"



[tool.poetry.group.dev.dependencies]
pylint = "^3.3.6"
black = "^25.1.0"
pre-commit = "^4.2.0"
pytest-mock = "^3.14.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
